requirements.txt
# --- Resume parsing ---
# PyMuPDF==1.24.10
pdfplumber==0.11.4
python-docx==1.1.2

# --- NLP & ML ---
spacy==3.7.5
nltk==3.9.1
scikit-learn==1.5.2
sentence-transformers==3.1.1
transformers==4.45.2
torch==2.4.1
xgboost==2.1.1
numpy==1.26.4
pandas==2.2.3

# --- API & utils ---
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.27.2
loguru==0.7.2

# --- Dev / test ---
pytest==8.3.3
pytest-cov==5.0.0
black==24.10.0
ruff==0.6.9
mypy==1.11.2
jupyter==1.1.1
